% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CoutQc.R
\name{CoutQc}
\alias{CoutQc}
\title{Class CoutQc}
\description{
The \code{CoutQc} class facilitates the counting of medical resource use
across individuals. Using databases structured by the \code{\link[=coutqc_bd]{coutqc_bd()}} function,
it features processes to convert resource utilization data, calculate
inflated costs at a reference year, and produce both an aggregated database
and a summary of average actual costs.
}
\seealso{
\code{\link[=coutqc_bd]{coutqc_bd()}}, \code{\link[=data_aggregate]{data_aggregate()}}

vignette("Introduction_macro_CoutsQc", package = "CoutsQc")
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{unit_cost}}{Accessor for the list of unit cost conversion tables.}

    \item{\code{cpi}}{Accessor for the Consumer price index conversion table.}

    \item{\code{reference_year}}{Accessor for reference_year attribute.}

    \item{\code{period_duration}}{Accessor for period_duration attribute.}

    \item{\code{follow_interval}}{Accessor for follow_interval attribute.}

    \item{\code{filter_period_limit}}{Accessor for filter_period_limit attribute.}

    \item{\code{cohort}}{Accessor for cohort database.}

    \item{\code{public_insurance}}{Accessor for public insurance database.}

    \item{\code{options}}{Accessor for coutsqc options.}

    \item{\code{aug_data}}{Accessor for augmented databases.}

    \item{\code{period}}{Accessor for period attribute.}

    \item{\code{nominal_cost}}{Accessor for nominal costs list.}

    \item{\code{actual_cost}}{Accessor for actual costs list.}

    \item{\code{indicator}}{Accessor for indicator list.}

    \item{\code{by_aggregate}}{Accessor for by_aggregate.}

    \item{\code{out_cohort}}{Accessor for output cohort.}

    \item{\code{data_summary}}{Accessor for data_summary attribute.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CoutQc-initialize}{\code{CoutQc$new()}}
    \item \href{#method-CoutQc-print}{\code{CoutQc$print()}}
    \item \href{#method-CoutQc-summary}{\code{CoutQc$summary()}}
    \item \href{#method-CoutQc-write}{\code{CoutQc$write()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoutQc-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CoutQc-initialize}{}}}
\subsection{\code{CoutQc$new()}}{
  Creating a new \code{CoutQc} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoutQc$new(
  ...,
  cohort = NULL,
  public_insurance = NULL,
  alive = FALSE,
  eligible = FALSE,
  filter_index = FALSE,
  filter_cohort_id = FALSE,
  filter_period_limit = NULL,
  analysis = TRUE,
  by_aggregate = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Pairs \code{<name> = <CoutQc_bd>}, a database object of
class \code{CoutQc_bd}. See \code{\link[=coutqc_bd]{coutqc_bd()}} for specification on the database
object. The given \verb{<name>} helps to identify the database in the function
output.}
      \item{\code{cohort}}{A \code{data.frame} of the cohort database.}
      \item{\code{public_insurance}}{A \code{data.frame} of individual's eligibility to
Quebec public insurance plan.}
      \item{\code{alive}}{\code{logical}. If \code{TRUE}, compute if an individual is
alive or dead at each periods.}
      \item{\code{eligible}}{\code{logical}. IF \code{TRUE}, compute if an individual is
covered or not by the Quebec public drug insurance plan at each periods.}
      \item{\code{filter_index}}{\code{logical}. If \code{TRUE}, exclude services
administered before an index date.}
      \item{\code{filter_cohort_id}}{\code{logical}. If \code{TRUE}, retaining only
individuals whose unique identifier \code{id} is present in the cohort database.}
      \item{\code{filter_period_limit}}{A vector of length 2 for the minimum and maximum
periods. It allows to specify a range of periods to keep. Values must be in
the range of those computed by the \code{coutqc_bd} period to work properly and
may change according to the \code{period_duration} argument. By default
\code{NULL}, keeps all observed periods.}
      \item{\code{analysis}}{\code{logical}; \code{TRUE}, performs analyses of augmented
databases. \code{FALSE}, to prevent analyses.}
      \item{\code{by_aggregate}}{A vector of column names used in the data aggregation.
By default \code{c("database", "period")}, aggregates actual costs by
\code{period} (generated by \code{CoutQc_bd}) and \code{database}(the name
of the database).}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{CoutQc} class object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoutQc-print"></a>}}
\if{latex}{\out{\hypertarget{method-CoutQc-print}{}}}
\subsection{\code{CoutQc$print()}}{
  Print information of an \code{CoutQc} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoutQc$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Further arguments passed to or from other methods.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoutQc-summary"></a>}}
\if{latex}{\out{\hypertarget{method-CoutQc-summary}{}}}
\subsection{\code{CoutQc$summary()}}{
  Brief summary of results.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoutQc$summary(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Further arguments passed to or from other methods.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoutQc-write"></a>}}
\if{latex}{\out{\hypertarget{method-CoutQc-write}{}}}
\subsection{\code{CoutQc$write()}}{
  Exports augmented databases to .csv files. The file names correspond to
the database names provided to the function, followed by
"_augmented.csv".
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoutQc$write(path, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{path}}{Name of the folder to which the files will be exported.}
      \item{\code{...}}{Further arguments passed to \code{\link[data.table:fwrite]{data.table::fwrite()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

}
