% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CoutQc_bd.R
\name{CoutQc_bd}
\alias{CoutQc_bd}
\alias{CoutQc_bd_iru}
\title{Class CoutQc_bd}
\description{
The \code{coutqc_bd} function allows to instance a \code{CoutQc_bd} object
of a Quebec medico-administrative database. Class objects created from this
function are used by other procedures of the package, namely the
\code{CoutQc} class. Its role is to validate the database, inflate service
costs, and identify service periods.

The \code{CoutQc_bd_iru} class is an extension of the \code{CoutQc_bd}
structure that allows the conversion of resource utilizations into costs.
}
\seealso{
\code{\link[=get_actual_cost]{get_actual_cost()}}, \code{\link[=get_period]{get_period()}}, \code{\link[=get_unit_cost]{get_unit_cost()}}; functions used by the
class.
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{cpi}}{Accessor for the Consumer price index conversion table.}

    \item{\code{reference_year}}{Accessor for reference_year attribute.}

    \item{\code{period_duration}}{Accessor for period_duration attribute.}

    \item{\code{follow_interval}}{Accessor for follow_interval attribute.}

    \item{\code{nominal_cost}}{Accessor for nominal_cost attribute.}

    \item{\code{actual_cost}}{Accessor for actual_cost attribute.}

    \item{\code{period}}{Accessor for period attribute.}

    \item{\code{aug_data}}{Accessor for augmented database.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CoutQc_bd-initialize}{\code{CoutQc_bd$new()}}
    \item \href{#method-CoutQc_bd-print}{\code{CoutQc_bd$print()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoutQc_bd-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CoutQc_bd-initialize}{}}}
\subsection{\code{CoutQc_bd$new()}}{
  Creating a new \code{CoutQc_bd} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoutQc_bd$new(
  data,
  reference_year,
  cpi,
  period_duration = "year",
  follow_interval = NULL,
  cohort = NULL,
  validate = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{A \code{data.frame} of a medico-administrative database.
Its must minimally be composed of an identifier \code{id}, a date variable for
the date of services named \code{service_date} and a variable for the value of
those services named \code{service_cost}. This value can be either a direct cost
or a resource utilization indicator.}
      \item{\code{reference_year}}{An \code{integer} of the reference year to which the
cost should be inflated to.}
      \item{\code{cpi}}{A \code{data.frame}; the consumer price index table. Its format is
composed of two variables: \code{year_cpi} corresponding to the calendar years
and \code{cpi} corresponding to the consumer price index of those years.}
      \item{\code{period_duration}}{A \code{character} to specifies the duration of the
period that is identified. By default, \code{"year"}, the calendar year is
identify. Available duration are \code{"year"}, \code{"fiscal"},
\code{"month"}, \code{"quarter"}, \code{“follow”}. See details for
additional information on supported duration.}
      \item{\code{follow_interval}}{An \code{integer} of the number of days in follow-up
intervals.}
      \item{\code{cohort}}{A \code{data.frame} of the cohort database. Used to supply
the \code{“date_index”} variable. Should be composed of one line by
identifier \code{id}.}
      \item{\code{validate}}{A \code{logical}. If \code{TRUE}, Perform additional
validations of inputs.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{CoutQc_bd} class object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoutQc_bd-print"></a>}}
\if{latex}{\out{\hypertarget{method-CoutQc_bd-print}{}}}
\subsection{\code{CoutQc_bd$print()}}{
  Print information of an \code{CoutQc_bd} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoutQc_bd$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Further arguments passed to or from other methods.}
    }
    \if{html}{\out{</div>}}
  }
}

}
\section{Super class}{
\code{\link[CoutsQc:CoutQc_bd]{CoutQc_bd}} -> \code{CoutQc_bd_iru}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{unit_cost}}{Accessor for the unit cost conversion table.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CoutQc_bd_iru-initialize}{\code{CoutQc_bd_iru$new()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="CoutsQc" data-topic="CoutQc_bd" data-id="print"><a href='../../CoutsQc/html/CoutQc_bd.html#method-CoutQc_bd-print'><code>CoutQc_bd$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoutQc_bd_iru-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CoutQc_bd_iru-initialize}{}}}
\subsection{\code{CoutQc_bd_iru$new()}}{
  Creating a new \code{CoutQc_bd_iru} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CoutQc_bd_iru$new(
  data,
  reference_year,
  cpi,
  unit_cost,
  period_duration = "year",
  follow_interval = NULL,
  cohort = NULL,
  validate = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{A \code{data.frame} of a medico-administrative database.
Its must minimally be composed of an identifier \code{id}, a date variable for
the date of services named \code{service_date} and a variable for the value of
those services named \code{service_cost}. This value can be either a direct cost
or a resource utilization indicator.}
      \item{\code{reference_year}}{An \code{integer} of the reference year to which the
cost should be inflated to.}
      \item{\code{cpi}}{A \code{data.frame}; the consumer price index table. Its format is
composed of two variables: \code{year_cpi} corresponding to the calendar years
and \code{cpi} corresponding to the consumer price index of those years.}
      \item{\code{unit_cost}}{A \code{data.frame} of unit cost conversion table. The
format of the table consists of two variables: \code{year_unit_cost}
corresponding to the year at the end of the fiscal year and \code{unit_cost}
corresponding to the cost of one unit of resource utilized. \code{NULL}
takes \code{service_cost} as nominal cost.}
      \item{\code{period_duration}}{A \code{character} to specifies the duration of the
period that is identified. By default, \code{"year"}, the calendar year is
identify. Available duration are \code{"year"}, \code{"fiscal"},
\code{"month"}, \code{"quarter"}, \code{“follow”}. See details for
additional information on supported duration.}
      \item{\code{follow_interval}}{An \code{integer} of the number of days in follow-up
intervals.}
      \item{\code{cohort}}{A \code{data.frame} of the cohort database. Used to supply
index dates.}
      \item{\code{validate}}{A \code{logical}. If \code{TRUE}, Perform additional
validations of inputs.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{CoutQc_bd} class object.
  }
}

}
