% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coutqc.R
\name{coutqc}
\alias{coutqc}
\title{Macro CoutQc}
\usage{
coutqc(..., cohort = NULL, public_insurance = NULL, alive = FALSE,
 eligible = FALSE, filter_index = FALSE, filter_cohort_id = FALSE,
 filter_period_limit = NULL, analysis = TRUE, by_aggregate = NULL)
}
\arguments{
\item{...}{Pairs \code{<name> = <CoutQc_bd>}, a database object of
class \code{CoutQc_bd}. See \code{\link[=coutqc_bd]{coutqc_bd()}} for specification on the database
object. The given \verb{<name>} helps to identify the database in the function
output.}

\item{cohort}{A \code{data.frame} of the cohort database.}

\item{public_insurance}{A \code{data.frame} of individual's eligibility to
Quebec public insurance plan.}

\item{alive}{\code{logical}. If \code{TRUE}, compute if an individual is
alive or dead at each periods.}

\item{eligible}{\code{logical}. IF \code{TRUE}, compute if an individual is
covered or not by the Quebec public drug insurance plan at each periods.}

\item{filter_index}{\code{logical}. If \code{TRUE}, exclude services
administered before an index date.}

\item{filter_cohort_id}{\code{logical}. If \code{TRUE}, retaining only
individuals whose unique identifier \code{id} is present in the cohort database.}

\item{filter_period_limit}{A vector of length 2 for the minimum and maximum
periods. It allows to specify a range of periods to keep. Values must be in
the range of those computed by the \code{coutqc_bd} period to work properly and
may change according to the \code{period_duration} argument. By default
\code{NULL}, keeps all observed periods.}

\item{analysis}{\code{logical}; \code{TRUE}, performs analyses of augmented
databases. \code{FALSE}, to prevent analyses.}

\item{by_aggregate}{A vector of column names used in the data aggregation.
By default \code{c("database", "period")}, aggregates actual costs by
\code{period} (generated by \code{CoutQc_bd}) and \code{database}(the name
of the database).}
}
\description{
The \code{coutqc()} function allows to instance a \link{CoutQc} object with a set
of Quebec medico-administrative databases.

The \code{CoutQc} class facilitates the counting of medical resource use
across individuals. Using databases structured by the \code{\link[=coutqc_bd]{coutqc_bd()}} function,
it features processes to convert resource utilization data, calculate
inflated costs at a reference year, and produce both an aggregated database
and a summary of average actual costs.
}
