% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoutsQc-package.R
\docType{package}
\name{CoutsQc-package}
\alias{CoutsQc}
\alias{CoutsQc-package}
\title{CoutsQc: Calculation of costs from Quebec's medical-administrative databases}
\description{
A macro that facilitates the accounting of inflated costs for Quebec's medico-administrative databases such as medical services, pharmaceutical services, hospital stays, and emergency room visits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LaRue-S/Package-CoutQc}
  \item Report bugs at \url{https://github.com/LaRue-S/Package-CoutQc/issues}
}

}
\author{
\strong{Maintainer}: Simon LaRue \email{simon.larue@crchudequebec.ulaval.ca} (\href{https://orcid.org/0009-0009-2344-0184}{ORCID})

Authors:
\itemize{
  \item Simon LaRue \email{simon.larue@crchudequebec.ulaval.ca} (\href{https://orcid.org/0009-0009-2344-0184}{ORCID})
  \item Jason R Gertin \email{jason.guertin@fmed.ulaval.ca} (\href{https://orcid.org/0000-0003-1718-5307}{ORCID})
  \item Marzia Angela Cremona (\href{https://orcid.org/0000-0001-8899-7316}{ORCID})
  \item Maude Laberge (\href{https://orcid.org/0000-0003-1274-136X}{ORCID})
}

}
\keyword{internal}
