% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregate.R
\name{data_aggregate}
\alias{data_aggregate}
\title{Aggregate actual costs}
\usage{
data_aggregate(aug_data, y, by, expand = TRUE, include_id = NULL)
}
\arguments{
\item{aug_data}{A \code{list} of \code{data.frame}, the databases augmented
with the actual cost and factor(s) of interest for each medical services.}

\item{y}{A \code{character} to specify the column name of the response
variable (usualy \code{"actual_cost"}).}

\item{by}{A \code{character} vector to specify the column names of the
factor(s) of interest.}

\item{expand}{A \code{logical}; If \code{TRUE}, adds zero costs in output
cohort for individuals who have not incurred costs for the factor(s) of
interest.}

\item{include_id}{A \code{character} vector of individual's id to be force in
the output. Used to include individual without cost in any
medico-administrative database.}
}
\value{
A \code{data.frame} of the aggregated costs for each individual.
}
\description{
This function aggregates a actual cost of medical services for each
individual by given factors.
}
