% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Summarize actual costs}
\usage{
data_summary(out_cohort, y, by)
}
\arguments{
\item{out_cohort}{A \code{data.frame} of a databases with the actual cost and
factor of interest for each id.}

\item{y}{A \code{character} to specify the column name of the response
variable (usualy \code{"actual_cost"}).}

\item{by}{A \code{character} vector to specify the column names of the
factor(s) of interest.}
}
\value{
A \code{list} of an overview of actual costs according to \code{by}
variables.
}
\description{
This function produce a small summary of actual cost according to given
factors.
}
