% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_demo.R
\name{generate_demo}
\alias{generate_demo}
\title{Demo Databases}
\usage{
generate_demo()
}
\value{
list of 6 demo datasets:
\itemize{
\item pharma_coutsqc, pharmaceutical services;
\item medical_coutsqc, medical services;
\item hospital_coutsqc, hospital services;
\item emergency_coutsqc, emergency visits;
\item insurance_coutsqc, RAMQ medical insurance eligibility;
\item cohort_coutsqc, cohort data.
}
}
\description{
Databases used for CoutsQc package demo
}
