% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actual_cost.R
\name{get_actual_cost}
\alias{get_actual_cost}
\title{Adjust nominal costs into actual costs}
\usage{
get_actual_cost(nominal_cost, year, reference_year, cpi)
}
\arguments{
\item{nominal_cost}{A \code{numeric} vector of nominal cost.}

\item{year}{A \code{integer} vector of service's calendar year.}

\item{reference_year}{\code{integer}. Reference year to adjust to.}

\item{cpi}{A \code{list} or \code{data.frame} of consumer price index
conversion table. Must be comprised of the variables named "year_cpi" and
"cpi".}
}
\value{
Vector of actual cost the same length as \code{nominal_cost}.
}
\description{
This function adjusts the nominal cost of a service according to the
consumer price index (CPI) for a given reference year.
}
\examples{
nominal_cost <- c(23.45, 12.99, 21.48, 34.00)
year <- c(2009, 2012, 2013, 2013)
reference_year <- 2013
cpi = data.frame(
  year_cpi = c(2009, 2010, 2011, 2012, 2013),
  cpi = c(113.4, 114.8, 118.3, 120.8, 121.7)
)
get_actual_cost(nominal_cost, year, reference_year, cpi)

}
\seealso{
\link{tbl_cpi}, Consumer price index conversion table.
}
