% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_period.R
\name{get_period}
\alias{get_period}
\title{Identify service periods}
\usage{
get_period(date, period_duration = c("year", "fiscal", "month",
 "quarter", "follow"), index_date = NULL, follow_interval = NULL)
}
\arguments{
\item{date}{A \code{Date} vector of service date.}

\item{period_duration}{A \code{character} string of the period duration.
\itemize{
\item \code{"year"} identifies the calendar year,
\item \code{"fiscal"} identifies the Quebec's fiscal year,
\item \code{"month"} identifies the month of the calendar year,
\item \code{"quarter"} identifies the quarter of the calendar year,
\item \code{"follow"} identifies follow-up period from an index date.
}}

\item{index_date}{A \code{Date} vector of index date the same length as
\code{date}.}

\item{follow_interval}{\code{numeric}. Length of days between follow-up
periods.}
}
\value{
Vector of period the same length as \code{date}.
}
\description{
This function identifies the period corresponding to a service date based on
a period duration.
}
\examples{
 # Calendar year as period duration.
 service_dates <- as.Date(c("2010-02-03", "2012-05-14", "2013-08-01"))
 get_period(service_dates, period_duration = "year")

 # 40-day follow-up period.
 service_dates <- as.Date(c("2011-02-03", "2010-12-14", "2011-05-01"))
 index_dates <- as.Date(c("2011-01-13", "2010-10-14", "2011-05-12"))
 get_period(
   service_dates,
   period_duration = "follow",
   index_date = index_dates,
   follow_interval = 40
 )

}
