% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_cpi}
\alias{tbl_cpi}
\title{Consumer Price Index, annual average, not seasonally adjusted, Quebec}
\format{
\code{Data.table} of two columns:
\describe{
\item{year_cpi}{Year}
\item{cpi}{Consumer price index}
}
}
\source{
Statistics Canada, Table 18-10-0005-01  Consumer Price Index, annual
average, not seasonally adjusted.
\url{https://doi.org/10.25318/1810000501-eng}
Last update: 2026-01-19
}
\usage{
tbl_cpi
}
\description{
Subset of data from the \emph{Institut de la statistique du Quebec} on the
consumer price index (CPI). The values are annual averages for all goods and
services (‘2002’ fixed at 100).
}
\keyword{datasets}
