% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_nirru}
\alias{tbl_nirru}
\title{Unit values data for hospital stays, Quebec}
\format{
\code{Data.table} of two columns:
\describe{
\item{year_unit_cost}{Fiscal year}
\item{unit_cost}{Unit value of one unit of NIRRU}
}
}
\source{
.
Last update: 2025
}
\usage{
tbl_nirru
}
\description{
Contains data from the Quebec Ministry of Health and Social Services on the
unit costs of level of relative intensity of resources used (known as NIRRU)
per fiscal period.
}
\keyword{datasets}
