% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_nirruc1j}
\alias{tbl_nirruc1j}
\title{Unit values data for for same-day surgery hospital stays, Quebec}
\format{
\code{Data.table} of two columns:
\describe{
\item{year_unit_cost}{Fiscal year}
\item{unit_cost}{Unit value of one unit of NIRRUc1j}
}
}
\source{
.
Last update: 2025
}
\usage{
tbl_nirruc1j
}
\description{
Contains data from the Quebec Ministry of Health and Social Services on the
unit costs of level of relative intensity of resources used for same-day
surgery (known as NIRRUc1j) per fiscal period.
}
\keyword{datasets}
