% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_urgence}
\alias{tbl_urgence}
\title{Emergency room visit unit value data, Quebec}
\format{
\code{Data.table} of two columns:
\describe{
\item{year_unit_cost}{Fiscal year}
\item{unit_cost}{Unit value of one emergency room visit}
}
}
\source{
MSSS.
Last update: 2026
}
\usage{
tbl_urgence
}
\description{
Contains data from Quebec's Ministry of Health and Social Services on unit
costs for an emergency room visit (regardless of duration) by fiscal period.
}
\keyword{datasets}
